make toggle non-reactive if there's no network
While it is possible to disable hblock without network access, it is not
possible to enable it back again (it requires internet access to
download its blocklists). That being said, I think it makes for a better
user experience to make the toggle non-sensitive when there's no network
to avoid the case that there's no way to undo an action (disabling
Blocker while offline, then not being able to enable it again).